requirements.txt
docxtpl
openai
python-dotenv
pandas
openpyxl